+++ docs/make.jl
@@ -16,7 +16,9 @@ makedocs(
              "Pauli Particles in 3D"         => "pauli_particle_3d.md",
              "Guiding Center Dynamics in 3D" => "guiding_center_3d.md",
              "Guiding Center Dynamics in 4D" => "guiding_center_4d.md",
+             "Gyrokinetics in 4D"            => "gyro_kinetics_4d.md",
              "Model Audit"                   => "audit.md",
+             "Findings"                      => "findings.md",
              "Examples" => [
                     "ITER Equilibrium in Cylindrical Coordinates"   => "examples/iter_cylindrical.md",
                  ],